RealmActivityMonitor/requirements.txt
requests>=2.0.0
beautifulsoup4>=4.9.0
boto3>=1.26.0  # For AWS SDK to interact with DynamoDB
python-dateutil